feat: restructure admin components and update import paths
- Refactored admin components to align with the new directory structure under `features/admin`.
- Updated import paths for `AccesDenied`, `AdminNav`, `SummaryView`, `CreateProductView`, `ManageProductsView`, and `ManageUsersView`.
- Introduced new views and components for better organization and functionality within the admin section.
- Enhanced hooks for product and user management to improve data handling and user experience.
