+++ app/admin/(adminPages)/manage-products/page.tsx
@@ -1,5 +1,5 @@
-import ManageProductsClient from "./ManageProductsClient";
+import ManageProductsView from "@/features/admin/views/ManageProductsView";
 
 export default async function page() {
-  return <ManageProductsClient />;
+  return <ManageProductsView />;
 }
