+++ app/admin/page.tsx
@@ -1,5 +1,5 @@
-import Summary from "./Summary";
+import SummaryView from "@/features/admin/views/SummaryView";
 
 export default async function page() {
-  return <Summary />;
+  return <SummaryView />;
 }
