+++ app/orders/Orders.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import ActionBtn from "@/components/shared/admin/ActionBtn";
 import PageState from "@/components/shared/PageState";
 import Title from "@/components/shared/Title";
 import { Card, CardContent } from "@/components/ui/card";
@@ -12,6 +11,7 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
+import ActionBtn from "@/features/admin/components/ActionBtn";
 import {
   useDeleteOrderForCurrentUser,
   useOrderForCurrentUser,
