+++ components/shared/sections/Collection.tsx
@@ -2,9 +2,9 @@
 
 import ProductCard from "@/components/shared/ProductCard";
 import Title from "@/components/shared/Title";
+import { useProduct } from "@/features/admin/hooks/useProduct";
 import { useCartActions } from "@/features/cart/hooks/useCartActions";
 import { useFavoriteActions } from "@/hooks/useFavoriteActions";
-import { useProduct } from "@/hooks/useProduct";
 import { useFavoriteStore } from "@/stores/FavoriteStore";
 
 export default function Collection() {
