+++ features/admin/components/ActionBtn.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { IconType } from "react-icons/lib";
-import { Button } from "../../ui/button";
+import { Button } from "../../../components/ui/button";
 
 type ActionBtnType = {
   icon: IconType;
